Squash watch edge cases
