+++ operator/src/main/java/oracle/kubernetes/operator/Main.java
@@ -5,25 +5,19 @@
 package oracle.kubernetes.operator;
 
 import io.kubernetes.client.JSON;
-import io.kubernetes.client.models.V1ConfigMap;
-import io.kubernetes.client.models.V1Event;
 import io.kubernetes.client.models.V1EventList;
-import io.kubernetes.client.models.V1ObjectMeta;
-import io.kubernetes.client.models.V1ObjectReference;
 import io.kubernetes.client.models.V1Pod;
 import io.kubernetes.client.models.V1PodList;
 import io.kubernetes.client.models.V1Service;
 import io.kubernetes.client.models.V1ServiceList;
 import io.kubernetes.client.models.V1beta1Ingress;
 import io.kubernetes.client.models.V1beta1IngressList;
-import io.kubernetes.client.util.Watch;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
@@ -31,13 +25,9 @@ import java.util.StringTokenizer;
 import java.util.TreeSet;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
-import oracle.kubernetes.operator.TuningParameters.MainTuning;
 import oracle.kubernetes.operator.calls.CallResponse;
 import oracle.kubernetes.operator.helpers.CRDHelper;
 import oracle.kubernetes.operator.helpers.CallBuilder;
@@ -47,39 +37,25 @@ import oracle.kubernetes.operator.helpers.DomainPresenceInfo;
 import oracle.kubernetes.operator.helpers.DomainPresenceInfoManager;
 import oracle.kubernetes.operator.helpers.HealthCheckHelper;
 import oracle.kubernetes.operator.helpers.HealthCheckHelper.KubernetesVersion;
-import oracle.kubernetes.operator.helpers.PodHelper;
 import oracle.kubernetes.operator.helpers.ResponseStep;
 import oracle.kubernetes.operator.helpers.ServerKubernetesObjects;
 import oracle.kubernetes.operator.helpers.ServerKubernetesObjectsManager;
-import oracle.kubernetes.operator.helpers.ServiceHelper;
-import oracle.kubernetes.operator.helpers.StorageHelper;
 import oracle.kubernetes.operator.logging.LoggingFacade;
 import oracle.kubernetes.operator.logging.LoggingFactory;
 import oracle.kubernetes.operator.logging.MessageKeys;
 import oracle.kubernetes.operator.rest.RestConfigImpl;
 import oracle.kubernetes.operator.rest.RestServer;
-import oracle.kubernetes.operator.steps.BeforeAdminServiceStep;
 import oracle.kubernetes.operator.steps.ConfigMapAfterStep;
-import oracle.kubernetes.operator.steps.DeleteDomainStep;
-import oracle.kubernetes.operator.steps.DomainPresenceStep;
-import oracle.kubernetes.operator.steps.ExternalAdminChannelsStep;
-import oracle.kubernetes.operator.steps.ListPersistentVolumeClaimStep;
-import oracle.kubernetes.operator.steps.ManagedServersUpStep;
-import oracle.kubernetes.operator.steps.WatchPodReadyAdminStep;
-import oracle.kubernetes.operator.wlsconfig.WlsRetriever;
 import oracle.kubernetes.operator.work.Component;
 import oracle.kubernetes.operator.work.Container;
 import oracle.kubernetes.operator.work.Engine;
-import oracle.kubernetes.operator.work.Fiber;
 import oracle.kubernetes.operator.work.Fiber.CompletionCallback;
-import oracle.kubernetes.operator.work.FiberGate;
 import oracle.kubernetes.operator.work.NextAction;
 import oracle.kubernetes.operator.work.Packet;
 import oracle.kubernetes.operator.work.Step;
 import oracle.kubernetes.operator.work.ThreadFactorySingleton;
 import oracle.kubernetes.weblogic.domain.v2.Domain;
 import oracle.kubernetes.weblogic.domain.v2.DomainList;
-import oracle.kubernetes.weblogic.domain.v2.DomainSpec;
 
 /** A Kubernetes Operator for WebLogic. */
 public class Main {
@@ -90,7 +66,7 @@ public class Main {
 
   private static final LoggingFacade LOGGER = LoggingFactory.getLogger("Operator", "Operator");
 
-  private static final TuningParameters tuningAndConfig;
+  static final TuningParameters tuningAndConfig;
 
   static {
     try {
@@ -123,17 +99,17 @@ public class Main {
                 callBuilderFactory));
   }
 
-  private static final Engine engine = new Engine(wrappedExecutorService);
-  private static final FiberGate FIBER_GATE = new FiberGate(engine);
+  static final Engine engine = new Engine(wrappedExecutorService);
 
-  private static final Map<String, AtomicBoolean> isNamespaceStopping = new ConcurrentHashMap<>();
+  static final Map<String, AtomicBoolean> isNamespaceStopping = new ConcurrentHashMap<>();
 
   private static final Map<String, ConfigMapWatcher> configMapWatchers = new ConcurrentHashMap<>();
   private static final Map<String, DomainWatcher> domainWatchers = new ConcurrentHashMap<>();
-  private static final Map<String, PodWatcher> podWatchers = new ConcurrentHashMap<>();
   private static final Map<String, EventWatcher> eventWatchers = new ConcurrentHashMap<>();
   private static final Map<String, ServiceWatcher> serviceWatchers = new ConcurrentHashMap<>();
   private static final Map<String, IngressWatcher> ingressWatchers = new ConcurrentHashMap<>();
+
+  static final Map<String, PodWatcher> podWatchers = new ConcurrentHashMap<>();
 
   private static final String operatorNamespace = getOperatorNamespace();
 
@@ -229,7 +205,7 @@ public class Main {
   }
 
   private static void completeBegin() {
-    deleteStrandedResources();
+    DomainProcessor.deleteStrandedResources();
 
     try {
       // start the REST server
@@ -245,6 +221,14 @@ public class Main {
     } catch (Throwable e) {
       LOGGER.warning(MessageKeys.EXCEPTION, e);
     }
+  }
+
+  static KubernetesVersion getVersion() {
+    return version;
+  }
+
+  static String getPrincipal() {
+    return principal;
   }
 
   private static class StartNamespacesStep extends Step {
@@ -301,28 +285,17 @@ public class Main {
     }
   }
 
-  static void deleteStrandedResources() {
-    for (Map.Entry<String, DomainPresenceInfo> entry :
-        DomainPresenceInfoManager.getDomainPresenceInfos().entrySet()) {
-      String domainUID = entry.getKey();
-      DomainPresenceInfo info = entry.getValue();
-      if (info != null && info.getDomain() == null) {
-        deleteDomainPresence(info, info.getNamespace(), domainUID);
-      }
-    }
-  }
-
   private static final AtomicBoolean UNINITIALIZED_NS_STOPPING = new AtomicBoolean(true);
 
-  private static AtomicBoolean isNamespaceStopping(String ns) {
+  static AtomicBoolean isNamespaceStopping(String ns) {
     return isNamespaceStopping.getOrDefault(ns, UNINITIALIZED_NS_STOPPING);
   }
 
-  private static void runSteps(Step firstStep) {
+  static void runSteps(Step firstStep) {
     runSteps(firstStep, null);
   }
 
-  private static void runSteps(Step firstStep, Runnable completionAction) {
+  static void runSteps(Step firstStep, Runnable completionAction) {
     engine.createFiber().start(firstStep, new Packet(), andThenDo(completionAction));
   }
 
@@ -339,7 +312,7 @@ public class Main {
       namespacesToStop.removeAll(targetNamespaces);
       stopNamespaces(namespacesToStop);
 
-      runSteps(new StartNamespacesStep(targetNamespaces), Main::deleteStrandedResources);
+      runSteps(new StartNamespacesStep(targetNamespaces), DomainProcessor::deleteStrandedResources);
     };
   }
 
@@ -385,7 +358,7 @@ public class Main {
 
   private static ConfigMapAfterStep createConfigMapStep(String ns) {
     return new ConfigMapAfterStep(
-        ns, configMapWatchers, isNamespaceStopping(ns), Main::dispatchConfigMapWatch);
+        ns, configMapWatchers, isNamespaceStopping(ns), DomainProcessor::dispatchConfigMapWatch);
   }
 
   // -----------------------------------------------------------------------------
@@ -394,345 +367,6 @@ public class Main {
   // after watch events are received.
   //
   // -----------------------------------------------------------------------------
-
-  /**
-   * Restarts the admin server, if already running
-   *
-   * @param principal Service principal
-   * @param domainUID Domain UID
-   */
-  public static void doRestartAdmin(String principal, String domainUID) {
-    DomainPresenceInfo info = DomainPresenceInfoManager.lookup(domainUID);
-    if (info != null) {
-      Domain dom = info.getDomain();
-      if (dom != null) {
-        doCheckAndCreateDomainPresence(dom, false, true, null, null, false);
-      }
-    }
-  }
-
-  /**
-   * Restarts the listed servers, if already running. Singleton servers will be immediately
-   * restarted. Clustered servers will be rolled so that the cluster maintains minimal availability,
-   * if possible.
-   *
-   * @param principal Service principal
-   * @param domainUID Domain UID
-   * @param servers Servers to roll
-   */
-  public static void doRollingRestartServers(
-      String principal, String domainUID, List<String> servers) {
-    DomainPresenceInfo info = DomainPresenceInfoManager.lookup(domainUID);
-    if (info != null) {
-      Domain dom = info.getDomain();
-      if (dom != null) {
-        doCheckAndCreateDomainPresence(dom, false, false, servers, null, false);
-      }
-    }
-  }
-
-  /**
-   * Restarts the listed clusters, if member servers are running. Member servers will be restarted
-   * in a rolling fashion in order to maintain minimal availability, if possible.
-   *
-   * @param principal Service principal
-   * @param domainUID Domain UID
-   * @param clusters Clusters to roll
-   */
-  public static void doRollingRestartClusters(
-      String principal, String domainUID, List<String> clusters) {
-    DomainPresenceInfo info = DomainPresenceInfoManager.lookup(domainUID);
-    if (info != null) {
-      Domain dom = info.getDomain();
-      if (dom != null) {
-        doCheckAndCreateDomainPresence(dom, false, false, null, clusters, false);
-      }
-    }
-  }
-
-  private static void scheduleDomainStatusUpdating(DomainPresenceInfo info) {
-    AtomicInteger unchangedCount = new AtomicInteger(0);
-    AtomicReference<ScheduledFuture<?>> statusUpdater = info.getStatusUpdater();
-    Runnable command =
-        new Runnable() {
-          public void run() {
-            try {
-              Runnable r = this; // resolve visibility
-              Packet packet = new Packet();
-              packet
-                  .getComponents()
-                  .put(
-                      ProcessingConstants.DOMAIN_COMPONENT_NAME,
-                      Component.createFor(info, version));
-              MainTuning main = tuningAndConfig.getMainTuning();
-              Step strategy =
-                  DomainStatusUpdater.createStatusStep(main.statusUpdateTimeoutSeconds, null);
-              engine
-                  .createFiber()
-                  .start(
-                      strategy,
-                      packet,
-                      new CompletionCallback() {
-                        @Override
-                        public void onCompletion(Packet packet) {
-                          Boolean isStatusUnchanged =
-                              (Boolean) packet.get(ProcessingConstants.STATUS_UNCHANGED);
-                          ScheduledFuture<?> existing = null;
-                          if (Boolean.TRUE.equals(isStatusUnchanged)) {
-                            if (unchangedCount.incrementAndGet()
-                                == main.unchangedCountToDelayStatusRecheck) {
-                              // slow down retries because of sufficient unchanged statuses
-                              existing =
-                                  statusUpdater.getAndSet(
-                                      engine
-                                          .getExecutor()
-                                          .scheduleWithFixedDelay(
-                                              r,
-                                              main.eventualLongDelay,
-                                              main.eventualLongDelay,
-                                              TimeUnit.SECONDS));
-                            }
-                          } else {
-                            // reset to trying after shorter delay because of changed status
-                            unchangedCount.set(0);
-                            existing =
-                                statusUpdater.getAndSet(
-                                    engine
-                                        .getExecutor()
-                                        .scheduleWithFixedDelay(
-                                            r,
-                                            main.initialShortDelay,
-                                            main.initialShortDelay,
-                                            TimeUnit.SECONDS));
-                            if (existing != null) {
-                              existing.cancel(false);
-                            }
-                          }
-                          if (existing != null) {
-                            existing.cancel(false);
-                          }
-                        }
-
-                        @Override
-                        public void onThrowable(Packet packet, Throwable throwable) {
-                          LOGGER.severe(MessageKeys.EXCEPTION, throwable);
-                          // retry to trying after shorter delay because of exception
-                          unchangedCount.set(0);
-                          ScheduledFuture<?> existing =
-                              statusUpdater.getAndSet(
-                                  engine
-                                      .getExecutor()
-                                      .scheduleWithFixedDelay(
-                                          r,
-                                          main.initialShortDelay,
-                                          main.initialShortDelay,
-                                          TimeUnit.SECONDS));
-                          if (existing != null) {
-                            existing.cancel(false);
-                          }
-                        }
-                      });
-            } catch (Throwable t) {
-              LOGGER.severe(MessageKeys.EXCEPTION, t);
-            }
-          }
-        };
-
-    MainTuning main = tuningAndConfig.getMainTuning();
-    ScheduledFuture<?> existing =
-        statusUpdater.getAndSet(
-            engine
-                .getExecutor()
-                .scheduleWithFixedDelay(
-                    command, main.initialShortDelay, main.initialShortDelay, TimeUnit.SECONDS));
-
-    if (existing != null) {
-      existing.cancel(false);
-    }
-  }
-
-  private static void doCheckAndCreateDomainPresence(Domain dom, boolean isWillInterrupt) {
-    doCheckAndCreateDomainPresence(dom, false, false, null, null, isWillInterrupt);
-  }
-
-  private static void doCheckAndCreateDomainPresence(
-      Domain dom,
-      boolean explicitRecheck,
-      boolean explicitRestartAdmin,
-      List<String> explicitRestartServers,
-      List<String> explicitRestartClusters,
-      boolean isWillInterrupt) {
-    LOGGER.entering();
-
-    boolean hasExplicitRestarts =
-        explicitRestartAdmin || explicitRestartServers != null || explicitRestartClusters != null;
-
-    DomainSpec spec = dom.getSpec();
-    DomainPresenceControl.normalizeDomainSpec(spec);
-    String domainUID = spec.getDomainUID();
-
-    DomainPresenceInfo existing = DomainPresenceInfoManager.lookup(domainUID);
-    if (existing != null) {
-      Domain current = existing.getDomain();
-      // Has the spec actually changed? We will get watch events for status updates
-      if (current != null
-          && (isOutdatedWatchEvent(current.getMetadata(), dom.getMetadata())
-              || (!explicitRecheck && !hasExplicitRestarts && spec.equals(current.getSpec())))) {
-        // nothing in the spec has changed or we received an outdated watch event
-        LOGGER.fine(MessageKeys.NOT_STARTING_DOMAINUID_THREAD, domainUID);
-        return;
-      }
-    }
-
-    DomainPresenceInfo info = DomainPresenceInfoManager.getOrCreate(dom);
-    info.setDomain(dom);
-
-    if (explicitRestartAdmin) {
-      LOGGER.info(MessageKeys.RESTART_ADMIN_STARTING, domainUID);
-      info.getExplicitRestartAdmin().set(true);
-    }
-    if (explicitRestartServers != null) {
-      LOGGER.info(MessageKeys.RESTART_SERVERS_STARTING, domainUID, explicitRestartServers);
-      info.getExplicitRestartServers().addAll(explicitRestartServers);
-    }
-    if (explicitRestartClusters != null) {
-      LOGGER.info(MessageKeys.ROLLING_CLUSTERS_STARTING, domainUID, explicitRestartClusters);
-      info.getExplicitRestartClusters().addAll(explicitRestartClusters);
-    }
-
-    checkAndCreateDomainPresence(info, isWillInterrupt);
-  }
-
-  private static void checkAndCreateDomainPresence(
-      DomainPresenceInfo info, boolean isWillInterrupt) {
-    Domain dom = info.getDomain();
-    DomainSpec spec = dom.getSpec();
-    String domainUID = spec.getDomainUID();
-
-    String ns = dom.getMetadata().getNamespace();
-    if (!isNamespaceStopping(ns).get()) {
-      LOGGER.info(MessageKeys.PROCESSING_DOMAIN, domainUID);
-      Step.StepAndPacket plan = createDomainUpPlan(info, ns);
-
-      CompletionCallback cc =
-          new CompletionCallback() {
-            @Override
-            public void onCompletion(Packet packet) {
-              info.complete();
-            }
-
-            @Override
-            public void onThrowable(Packet packet, Throwable throwable) {
-              LOGGER.severe(MessageKeys.EXCEPTION, throwable);
-
-              FIBER_GATE.startFiberIfLastFiberMatches(
-                  domainUID,
-                  Fiber.getCurrentIfSet(),
-                  DomainStatusUpdater.createFailedStep(throwable, null),
-                  plan.packet,
-                  new CompletionCallback() {
-                    @Override
-                    public void onCompletion(Packet packet) {
-                      // no-op
-                    }
-
-                    @Override
-                    public void onThrowable(Packet packet, Throwable throwable) {
-                      LOGGER.severe(MessageKeys.EXCEPTION, throwable);
-                    }
-                  });
-
-              FIBER_GATE
-                  .getExecutor()
-                  .schedule(
-                      () -> checkAndCreateDomainPresence(info, false),
-                      DomainPresence.getDomainPresenceFailureRetrySeconds(),
-                      TimeUnit.SECONDS);
-            }
-          };
-
-      if (isWillInterrupt) {
-        FIBER_GATE.startFiber(domainUID, plan.step, plan.packet, cc);
-      } else {
-        FIBER_GATE.startFiberIfNoCurrentFiber(domainUID, plan.step, plan.packet, cc);
-      }
-
-      scheduleDomainStatusUpdating(info);
-    }
-    LOGGER.exiting();
-  }
-
-  static Step.StepAndPacket createDomainUpPlan(DomainPresenceInfo info, String ns) {
-    PodWatcher pw = podWatchers.get(ns);
-    Step managedServerStrategy =
-        bringManagedServersUp(DomainStatusUpdater.createEndProgressingStep(null));
-    Step adminServerStrategy = bringAdminServerUp(info, managedServerStrategy);
-
-    Step strategy =
-        DomainStatusUpdater.createProgressingStep(
-            DomainStatusUpdater.INSPECTING_DOMAIN_PROGRESS_REASON,
-            true,
-            DomainPresenceStep.createDomainPresenceStep(
-                info, adminServerStrategy, managedServerStrategy));
-
-    Packet p = new Packet();
-    p.getComponents()
-        .put(ProcessingConstants.DOMAIN_COMPONENT_NAME, Component.createFor(info, version, pw));
-    p.put(ProcessingConstants.PRINCIPAL, principal);
-
-    return new Step.StepAndPacket(strategy, p);
-  }
-
-  // pre-conditions: DomainPresenceInfo SPI
-  // "principal"
-  private static Step bringAdminServerUp(DomainPresenceInfo info, Step next) {
-    return StorageHelper.insertStorageSteps(
-        info,
-        Step.chain(
-            new ListPersistentVolumeClaimStep(null),
-            PodHelper.createAdminPodStep(null),
-            new BeforeAdminServiceStep(null),
-            ServiceHelper.createForServerStep(null),
-            new WatchPodReadyAdminStep(podWatchers, null),
-            WlsRetriever.readConfigStep(null),
-            new ExternalAdminChannelsStep(next)));
-  }
-
-  private static Step bringManagedServersUp(Step next) {
-    return new ManagedServersUpStep(next);
-  }
-
-  static void deleteDomainPresence(Domain dom) {
-    String domainUID = dom.getSpec().getDomainUID();
-    DomainPresenceInfo existing = DomainPresenceInfoManager.lookup(domainUID);
-    if (existing != null) {
-      Domain current = existing.getDomain();
-      if (current != null && isOutdatedWatchEvent(current.getMetadata(), dom.getMetadata())) {
-        return;
-      }
-      existing.setDomain(null);
-    }
-    deleteDomainPresence(existing, dom.getMetadata().getNamespace(), domainUID);
-  }
-
-  private static void deleteDomainPresence(
-      DomainPresenceInfo info, String namespace, String domainUID) {
-    FIBER_GATE.startFiber(
-        domainUID,
-        new DeleteDomainStep(info, namespace, domainUID),
-        new Packet(),
-        new CompletionCallback() {
-          @Override
-          public void onCompletion(Packet packet) {
-            // no-op
-          }
-
-          @Override
-          public void onThrowable(Packet packet, Throwable throwable) {
-            LOGGER.severe(MessageKeys.EXCEPTION, throwable);
-          }
-        });
-  }
 
   /**
    * Obtain the list of target namespaces
@@ -795,40 +429,8 @@ public class Main {
         ns,
         READINESS_PROBE_FAILURE_EVENT_FILTER,
         initialResourceVersion,
-        Main::dispatchEventWatch,
+        DomainProcessor::dispatchEventWatch,
         isNamespaceStopping(ns));
-  }
-
-  private static void dispatchEventWatch(Watch.Response<V1Event> item) {
-    V1Event e = item.object;
-    if (e != null) {
-      switch (item.type) {
-        case "ADDED":
-        case "MODIFIED":
-          onEvent(e);
-          break;
-        case "DELETED":
-        case "ERROR":
-        default:
-      }
-    }
-  }
-
-  private static void onEvent(V1Event event) {
-    V1ObjectReference ref = event.getInvolvedObject();
-    if (ref != null) {
-      String name = ref.getName();
-      String message = event.getMessage();
-      if (message != null) {
-        if (message.contains(WebLogicConstants.READINESS_PROBE_NOT_READY_STATE)) {
-          ServerKubernetesObjects sko = ServerKubernetesObjectsManager.lookup(name);
-          if (sko != null) {
-            int idx = message.lastIndexOf(':');
-            sko.getLastKnownStatus().set(message.substring(idx + 1).trim());
-          }
-        }
-      }
-    }
   }
 
   private static PodWatcher createPodWatcher(String ns, String initialResourceVersion) {
@@ -836,52 +438,8 @@ public class Main {
         getThreadFactory(),
         ns,
         initialResourceVersion,
-        Main::dispatchPodWatch,
+        DomainProcessor::dispatchPodWatch,
         isNamespaceStopping(ns));
-  }
-
-  private static void dispatchPodWatch(Watch.Response<V1Pod> item) {
-    V1Pod p = item.object;
-    if (p != null) {
-      V1ObjectMeta metadata = p.getMetadata();
-      String domainUID = metadata.getLabels().get(LabelConstants.DOMAINUID_LABEL);
-      String serverName = metadata.getLabels().get(LabelConstants.SERVERNAME_LABEL);
-      if (domainUID != null) {
-        DomainPresenceInfo info = DomainPresenceInfoManager.lookup(domainUID);
-        if (info != null && serverName != null) {
-          ServerKubernetesObjects sko =
-              ServerKubernetesObjectsManager.getOrCreate(info, domainUID, serverName);
-          if (sko != null) {
-            switch (item.type) {
-              case "ADDED":
-                sko.getPod().set(p);
-                break;
-              case "MODIFIED":
-                V1Pod skoPod = sko.getPod().get();
-                if (skoPod != null) {
-                  // If the skoPod is null then the operator deleted this pod
-                  // and modifications are to the terminating pod
-                  sko.getPod().compareAndSet(skoPod, p);
-                }
-                break;
-              case "DELETED":
-                sko.getLastKnownStatus().set(WebLogicConstants.SHUTDOWN_STATE);
-                V1Pod oldPod = sko.getPod().getAndSet(null);
-                if (oldPod != null) {
-                  // Pod was deleted, but sko still contained a non-null entry
-                  LOGGER.info(
-                      MessageKeys.POD_DELETED, domainUID, metadata.getNamespace(), serverName);
-                  doCheckAndCreateDomainPresence(info.getDomain(), true);
-                }
-                break;
-
-              case "ERROR":
-              default:
-            }
-          }
-        }
-      }
-    }
   }
 
   private static ServiceWatcher createServiceWatcher(String ns, String initialResourceVersion) {
@@ -889,102 +447,8 @@ public class Main {
         getThreadFactory(),
         ns,
         initialResourceVersion,
-        Main::dispatchServiceWatch,
+        DomainProcessor::dispatchServiceWatch,
         isNamespaceStopping(ns));
-  }
-
-  private static void dispatchServiceWatch(Watch.Response<V1Service> item) {
-    V1Service s = item.object;
-    if (s != null) {
-      V1ObjectMeta metadata = s.getMetadata();
-      String domainUID = metadata.getLabels().get(LabelConstants.DOMAINUID_LABEL);
-      String serverName = metadata.getLabels().get(LabelConstants.SERVERNAME_LABEL);
-      String channelName = metadata.getLabels().get(LabelConstants.CHANNELNAME_LABEL);
-      String clusterName = metadata.getLabels().get(LabelConstants.CLUSTERNAME_LABEL);
-      if (domainUID != null) {
-        DomainPresenceInfo info = DomainPresenceInfoManager.lookup(domainUID);
-        ServerKubernetesObjects sko = null;
-        if (info != null) {
-          if (serverName != null) {
-            sko = ServerKubernetesObjectsManager.getOrCreate(info, domainUID, serverName);
-          }
-          switch (item.type) {
-            case "ADDED":
-              if (sko != null) {
-                if (channelName != null) {
-                  sko.getChannels().put(channelName, s);
-                } else {
-                  sko.getService().set(s);
-                }
-              } else if (clusterName != null) {
-                info.getClusters().put(clusterName, s);
-              }
-              break;
-            case "MODIFIED":
-              if (sko != null) {
-                if (channelName != null) {
-                  V1Service skoService = sko.getChannels().get(channelName);
-                  if (skoService != null) {
-                    sko.getChannels().replace(channelName, skoService, s);
-                  }
-                } else {
-                  V1Service skoService = sko.getService().get();
-                  if (skoService != null) {
-                    sko.getService().compareAndSet(skoService, s);
-                  }
-                }
-              } else if (clusterName != null) {
-                V1Service clusterService = info.getClusters().get(clusterName);
-                if (clusterService != null) {
-                  info.getClusters().replace(clusterName, clusterService, s);
-                }
-              }
-              break;
-            case "DELETED":
-              if (sko != null) {
-                if (channelName != null) {
-                  V1Service oldService = sko.getChannels().remove(channelName);
-                  if (oldService != null) {
-                    // Service was deleted, but sko still contained a non-null entry
-                    LOGGER.info(
-                        MessageKeys.SERVER_SERVICE_DELETED,
-                        domainUID,
-                        metadata.getNamespace(),
-                        serverName);
-                    doCheckAndCreateDomainPresence(info.getDomain(), true);
-                  }
-                } else {
-                  V1Service oldService = sko.getService().getAndSet(null);
-                  if (oldService != null) {
-                    // Service was deleted, but sko still contained a non-null entry
-                    LOGGER.info(
-                        MessageKeys.SERVER_SERVICE_DELETED,
-                        domainUID,
-                        metadata.getNamespace(),
-                        serverName);
-                    doCheckAndCreateDomainPresence(info.getDomain(), true);
-                  }
-                }
-              } else if (clusterName != null) {
-                V1Service oldService = info.getClusters().remove(clusterName);
-                if (oldService != null) {
-                  // Service was deleted, but clusters still contained a non-null entry
-                  LOGGER.info(
-                      MessageKeys.CLUSTER_SERVICE_DELETED,
-                      domainUID,
-                      metadata.getNamespace(),
-                      clusterName);
-                  doCheckAndCreateDomainPresence(info.getDomain(), true);
-                }
-              }
-              break;
-
-            case "ERROR":
-            default:
-          }
-        }
-      }
-    }
   }
 
   private static IngressWatcher createIngressWatcher(String ns, String initialResourceVersion) {
@@ -992,107 +456,11 @@ public class Main {
         getThreadFactory(),
         ns,
         initialResourceVersion,
-        Main::dispatchIngressWatch,
+        DomainProcessor::dispatchIngressWatch,
         isNamespaceStopping(ns));
   }
 
-  private static void dispatchIngressWatch(Watch.Response<V1beta1Ingress> item) {
-    V1beta1Ingress i = item.object;
-    if (i != null) {
-      V1ObjectMeta metadata = i.getMetadata();
-      String domainUID = metadata.getLabels().get(LabelConstants.DOMAINUID_LABEL);
-      String clusterName = metadata.getLabels().get(LabelConstants.CLUSTERNAME_LABEL);
-      if (domainUID != null) {
-        DomainPresenceInfo info = DomainPresenceInfoManager.lookup(domainUID);
-        if (info != null && clusterName != null) {
-          switch (item.type) {
-            case "ADDED":
-              info.getIngresses().put(clusterName, i);
-              break;
-            case "MODIFIED":
-              V1beta1Ingress skoIngress = info.getIngresses().get(clusterName);
-              if (skoIngress != null) {
-                info.getIngresses().replace(clusterName, skoIngress, i);
-              }
-              break;
-            case "DELETED":
-              V1beta1Ingress oldIngress = info.getIngresses().remove(clusterName);
-              if (oldIngress != null) {
-                // Ingress was deleted, but sko still contained a non-null entry
-                LOGGER.info(
-                    MessageKeys.INGRESS_DELETED, domainUID, metadata.getNamespace(), clusterName);
-                doCheckAndCreateDomainPresence(info.getDomain(), true);
-              }
-              break;
-
-            case "ERROR":
-            default:
-          }
-        }
-      }
-    }
-  }
-
-  private static void dispatchConfigMapWatch(Watch.Response<V1ConfigMap> item) {
-    V1ConfigMap c = item.object;
-    if (c != null) {
-      switch (item.type) {
-        case "MODIFIED":
-        case "DELETED":
-          runSteps(
-              ConfigMapHelper.createScriptConfigMapStep(
-                  getOperatorNamespace(), c.getMetadata().getNamespace()));
-          break;
-
-        case "ERROR":
-        default:
-      }
-    }
-  }
-
-  /**
-   * Dispatch the Domain event to the appropriate handler.
-   *
-   * @param item An item received from a Watch response.
-   */
-  private static void dispatchDomainWatch(Watch.Response<Domain> item) {
-    Domain d;
-    String domainUID;
-    switch (item.type) {
-      case "ADDED":
-      case "MODIFIED":
-        d = item.object;
-        domainUID = d.getSpec().getDomainUID();
-        LOGGER.info(MessageKeys.WATCH_DOMAIN, domainUID);
-        doCheckAndCreateDomainPresence(d, true);
-        break;
-
-      case "DELETED":
-        d = item.object;
-        domainUID = d.getSpec().getDomainUID();
-        LOGGER.info(MessageKeys.WATCH_DOMAIN_DELETED, domainUID);
-        deleteDomainPresence(d);
-        break;
-
-      case "ERROR":
-      default:
-    }
-  }
-
-  /* Recently, we've seen a number of intermittent bugs where K8s reports
-   * outdated watch events.  There seem to be two main cases: 1) a DELETED
-   * event for a resource that was deleted, but has since been recreated, and 2)
-   * a MODIFIED event for an object that has already had subsequent modifications.
-   */
-  static boolean isOutdatedWatchEvent(V1ObjectMeta current, V1ObjectMeta ob) {
-    return ob == null
-        || (current != null
-            && (Integer.parseInt(ob.getResourceVersion())
-                    < Integer.parseInt(current.getResourceVersion())
-                || ob.getCreationTimestamp().isBefore(current.getCreationTimestamp())));
-  }
-
-  private static String getOperatorNamespace() {
+  static String getOperatorNamespace() {
     String namespace = System.getenv("OPERATOR_NAMESPACE");
     if (namespace == null) {
       namespace = "default";
@@ -1164,8 +532,11 @@ public class Main {
       Set<String> domainUIDs = new HashSet<>();
       if (callResponse.getResult() != null) {
         for (Domain dom : callResponse.getResult().getItems()) {
-          domainUIDs.add(dom.getSpec().getDomainUID());
-          doCheckAndCreateDomainPresence(dom, false);
+          String domainUID = dom.getSpec().getDomainUID();
+          domainUIDs.add(domainUID);
+          DomainPresenceInfo info = DomainPresenceInfoManager.getOrCreate(dom);
+          info.setDomain(dom);
+          DomainProcessor.makeRightDomainPresence(info, dom, false, false);
         }
       }
 
@@ -1198,7 +569,7 @@ public class Main {
           getThreadFactory(),
           ns,
           initialResourceVersion,
-          Main::dispatchDomainWatch,
+          DomainProcessor::dispatchDomainWatch,
           isNamespaceStopping(ns));
     }
   }
@@ -1265,11 +636,8 @@ public class Main {
     @Override
     public NextAction onSuccess(Packet packet, CallResponse<V1EventList> callResponse) {
       V1EventList result = callResponse.getResult();
-      if (result != null) {
-        for (V1Event event : result.getItems()) {
-          onEvent(event);
-        }
-      }
+      // don't bother processing pre-existing events
+
       if (!eventWatchers.containsKey(ns)) {
         eventWatchers.put(ns, createEventWatcher(ns, getInitialResourceVersion(result)));
       }
