+++ operator/src/main/java/oracle/kubernetes/operator/helpers/DomainPresenceInfo.java
@@ -15,9 +15,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.CopyOnWriteArraySet;
 import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 import oracle.kubernetes.operator.wlsconfig.WlsDomainConfig;
 import oracle.kubernetes.operator.wlsconfig.WlsServerConfig;
@@ -39,10 +37,6 @@ public class DomainPresenceInfo {
   private final ConcurrentMap<String, ServerKubernetesObjects> servers = new ServerMap();
   private final ConcurrentMap<String, V1Service> clusters = new ConcurrentHashMap<>();
   private final ConcurrentMap<String, V1beta1Ingress> ingresses = new ConcurrentHashMap<>();
-
-  private final AtomicBoolean explicitRestartAdmin = new AtomicBoolean(false);
-  private final Set<String> explicitRestartServers = new CopyOnWriteArraySet<>();
-  private final Set<String> explicitRestartClusters = new CopyOnWriteArraySet<>();
 
   private V1PersistentVolumeClaimList claims = null;
 
@@ -200,33 +194,6 @@ public class DomainPresenceInfo {
    */
   public ConcurrentMap<String, V1beta1Ingress> getIngresses() {
     return ingresses;
-  }
-
-  /**
-   * Control for if domain has outstanding restart admin server pending
-   *
-   * @return Control for pending admin server restart
-   */
-  public AtomicBoolean getExplicitRestartAdmin() {
-    return explicitRestartAdmin;
-  }
-
-  /**
-   * Control list for outstanding server restarts
-   *
-   * @return Control list for outstanding server restarts
-   */
-  public Set<String> getExplicitRestartServers() {
-    return explicitRestartServers;
-  }
-
-  /**
-   * Control list for outstanding cluster restarts
-   *
-   * @return Control list for outstanding cluster restarts
-   */
-  public Set<String> getExplicitRestartClusters() {
-    return explicitRestartClusters;
   }
 
   /**
