+++ operator/src/main/java/oracle/kubernetes/operator/helpers/PodStepContext.java
@@ -33,8 +33,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.atomic.AtomicBoolean;
 import oracle.kubernetes.operator.KubernetesConstants;
 import oracle.kubernetes.operator.LabelConstants;
 import oracle.kubernetes.operator.PodAwaiterStepFactory;
@@ -283,20 +281,6 @@ public abstract class PodStepContext {
 
   abstract String getPodReplacedMessageKey();
 
-  AtomicBoolean getExplicitRestartAdmin() {
-    return info.getExplicitRestartAdmin();
-  }
-
-  abstract void updateRestartForNewPod();
-
-  abstract void updateRestartForReplacedPod();
-
-  Set<String> getExplicitRestartClusters() {
-    return info.getExplicitRestartClusters();
-  }
-
-  protected abstract boolean isExplicitRestartThisServer();
-
   Step createCyclePodStep(Step next) {
     return new CyclePodStep(next);
   }
@@ -315,7 +299,7 @@ public abstract class PodStepContext {
   }
 
   private boolean canUseCurrentPod(V1Pod currentPod) {
-    return !isExplicitRestartThisServer() && isCurrentPodValid(getPodModel(), currentPod);
+    return isCurrentPodValid(getPodModel(), currentPod);
   }
 
   // We want to detect changes that would require replacing an existing Pod
@@ -389,10 +373,6 @@ public abstract class PodStepContext {
     return false;
   }
 
-  Set<String> getExplicitRestartServers() {
-    return info.getExplicitRestartServers();
-  }
-
   private class VerifyPodStep extends Step {
 
     VerifyPodStep(Step next) {
@@ -403,7 +383,6 @@ public abstract class PodStepContext {
     public NextAction apply(Packet packet) {
       V1Pod currentPod = getSko().getPod().get();
       if (currentPod == null) {
-        updateRestartForNewPod();
         return doNext(createNewPod(getNext()), packet);
       } else if (canUseCurrentPod(currentPod)) {
         logPodExists();
@@ -457,7 +436,6 @@ public abstract class PodStepContext {
 
     @Override
     public NextAction onSuccess(Packet packet, CallResponse<V1Status> callResponses) {
-      updateRestartForReplacedPod();
       return doNext(replacePod(getNext()), packet);
     }
   }
